requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dnspython==2.7.0
gunicorn==22.0.0
kombu==5.5.4
mongoengine==0.29.1
packaging==25.0
prompt_toolkit==3.0.51
pymongo==4.13.0
python-dateutil==2.9.0.post0
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
telethon==1.33.1
tzdata==2025.2
uvicorn==0.30.3
vine==5.1.0
wcwidth==0.2.13
